ceph-fuse: test ability to remount on startup and fail out if we can't
Unprivileged users can't use options when remounting; see
http://tracker.ceph.com/issues/10542

We add a Client function that tries to invoke the remount function and
will return a failure if remounting is required. Then when starting
ceph-fuse we spin off another child thread to invoke this, and if it
fails we unmount the FS.

(Most unfortunately, there doesn't seem to be a good interface to tell
FUSE to shut down. See
http://fuse.996288.n3.nabble.com/libfuse-exiting-fuse-session-loop-td10686.html
I tried changing our signal invocation or attempting a simple action on
the mount point but those were ineffectual at terminating the remaining
processes; fusermount actually gets rid of them all.)

Signed-off-by: Greg Farnum <gfarnum@redhat.com>